adjust aot color scale

+++ cmaq.python/rzdm_plot_aot_aqmv710.py
@@ -15,7 +15,6 @@ import sys
 import datetime
 import shutil
 import subprocess
-
 user=os.environ['USER']
 
 script_dir=os.getcwd()
@@ -30,21 +29,25 @@ for line in rfile:
         ver=line.split("=")
         ver_name=ver[0].split(" ")
         if ver_name[1] == "aqm_ver":
-            aqm_ver_prod=ver[1]
+            aqm_ver=ver[1]
 rfile.close()
+if aqm_ver=="":
+    aqm_ver="v7.0"
+print("aqm_ver="+aqm_ver)
+
+wgrib2=os.environ['WGRIB2']
+if wgrib2 == "":
+    print("No definition of WGRIB2 can be found, please load module wgrib2/2.0.8")
+    sys.exit()
 
 ### PASSED AGRUEMENTS
-if len(sys.argv) < 2:
-    print("you must set 2 arguments as start_date end_date")
+if len(sys.argv) < 3:
+    print("you must set 3 arguments as cycle[06|12|all]  start_date end_date")
     sys.exit()
 else:
-    start_date = sys.argv[1]
-    end_date = sys.argv[2]
-
-comout="/lfs/h2/emc/physics/noscrub/"+os.environ['USER']+"/GOES16_AOD/REGRID"
-if not os.path.exists(comout):
-    print("Can not find output dir "+comout)
-    sys.exit()
+    sel_cyc = sys.argv[1]
+    start_date = sys.argv[2]
+    end_date = sys.argv[3]
 
 stmp_dir="/lfs/h2/emc/stmp/"+user
 if not os.path.exists(stmp_dir):
@@ -64,13 +67,13 @@ if nfind == -1:
     workid=py_code
 else:
     workid=py_code[0:nfind-1]
-working_dir=stmp_dir+"/"+workid
+working_dir=stmp_dir+"/aod_"+workid
 if not os.path.exists(working_dir):
     os.mkdir(working_dir)
 
 os.chdir(working_dir)
 
-msg_file=working_dir+"/msg_"+start_date
+msg_file=working_dir+"/msg_rzdm_aot_v7_"+start_date+"_"+sel_cyc
 cmd="cat /etc/cluster_name"
 subprocess.call([cmd+" > "+msg_file], shell=True)
 cmd="cat /etc/wcoss.conf | grep cluster_name | awk -F\":\" '{print $2}'"
@@ -82,10 +85,8 @@ if os.path.isfile(msg_file):
     sh.close()
 if machine.lower() == "dogwood":
     remote="cactus"
-    remote_host="clogin01.wcoss2.ncep.noaa.gov"
 elif machine.lower() == "cactus":
     remote="dogwood"
-    remote_host="dlogin01.wcoss2.ncep.noaa.gov"
 else:
     print("System name not defined for this script")
     sys.exit()
@@ -117,7 +118,6 @@ print(msg)
 
 sdate = datetime.datetime(int(start_date[0:4]), int(start_date[4:6]), int(start_date[6:]))
 edate = datetime.datetime(int(end_date[0:4]), int(end_date[4:6]), int(end_date[6:]))
-
 YMDH_date_format = "%Y%m%d/%H"
 YMD_date_format = "%Y%m%d"
 YM_date_format = "%Y%m"
@@ -128,24 +128,24 @@ H_date_format = "%H"
 date_inc = datetime.timedelta(hours=24)
 hour_inc = datetime.timedelta(hours=1)
 
-expid="aqm"
-expid="aqmv7"
+var=[ "aod" ]
+comout="/lfs/h1/ops/prod/com/aqm/"+aqm_ver
 
-warnings.filterwarnings('ignore')
-plt.rcParams['font.weight'] = 'bold'
-plt.rcParams['axes.labelsize'] = 10
-plt.rcParams['axes.labelweight'] = 'bold'
-plt.rcParams['xtick.labelsize'] = 10
-plt.rcParams['ytick.labelsize'] = 10
-plt.rcParams['axes.titlesize'] = 15
-plt.rcParams['axes.titleweight'] = 'bold'
-plt.rcParams['axes.formatter.useoffset'] = False
-## cbar_num_format = "%d"
-cbar_num_format = "%.2f"
-plt.close('all') # close all figures
+num_var=len(var)
+print("var length = "+str(num_var))
+if num_var == 0:
+    print(f"no variable to scp")
+    sys.exit()
 
-msg=datetime.datetime.now()
-msg=msg - date_inc
+if sel_cyc == "all":
+   cyc_opt=[ "06", "12" ]
+elif sel_cyc == "06":
+   cyc_opt=[ "06" ]
+elif sel_cyc == "12":
+   cyc_opt=[ "12" ]
+else:
+    print("seletced cycle"+sel_cyc+" can not be recongized.")
+    sys.exit()
 
 figout=stmp_dir
 
@@ -154,19 +154,18 @@ while date <= edate:
     YY=date.strftime(Y_date_format)
     YM=date.strftime(YM_date_format)
     YMD=date.strftime(YMD_date_format)
-    
-    figdir = figout+"/goes16_"+expid+"_"+YMD
-    print(figdir)
-
-    os.chdir(figdir)
-    parta=os.path.join("/usr", "bin", "scp")
-    if 1 == 1 :
-        partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "web", "fig", YY, YMD)
-    else:
-        partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "transfer")
-        partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "ftp")
-    subprocess.call(['scp -p * '+partb], shell=True)
-    print("FIG DIR = "+figdir)
-    msg=datetime.datetime.now()
-    print("End   processing "+YMD+" Current system time is :: "+msg.strftime("%Y-%m-%d %H:%M:%S"))
+    for cyc in cyc_opt:
+        cycle="t"+cyc+"z"
+        for ivar in range(0,num_var):
+            figdir = figout+"/aqm"+"_prod_"+YMD+"_"+var[ivar]+"_"+cycle
+            if os.path.exists(figdir):
+                os.chdir(figdir)
+                parta=os.path.join("/usr", "bin", "scp")
+                if 1 == 1 :
+                    partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "web", "fig", date.strftime(Y_date_format), YMD, cycle)
+                else:
+                    partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "ftp")
+                    partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "transfer")
+                subprocess.call(['scp -p * '+partb], shell=True)
+                print("FIG DIR = "+figdir)
     date = date + date_inc
